+++ Kernel/Time/RTC.cpp
@@ -6,6 +6,7 @@
 
 #include <Kernel/Arch/x86/IO.h>
 #include <Kernel/Arch/x86/InterruptDisabler.h>
+#include <Kernel/Arch/x86/NonMaskableInterruptDisabler.h>
 #include <Kernel/CMOS.h>
 #include <Kernel/Time/RTC.h>
 #include <Kernel/Time/TimeManagement.h>
